Tidy up: change to q... types, delete empty rows

+++ testlib/efentoenvironmentsensortest.h
@@ -10,7 +10,7 @@ public:
     void decodeTemperatureTest(const QByteArray &manufData, bool &valueChanged);
     void decodeHumidityTest(const QByteArray &manufData, bool &valueChanged);
     void decodeAirPressureTest(const QByteArray &manufData, bool &valueChanged);
-    static float zigzagConvertTest(unsigned long valueRaw, float divisor);
+    static float zigzagConvertTest(quint32 valueRaw, float divisor);
     void resetErrorFlagsTest();
